+++ perf/__init__.py
@@ -20,8 +20,8 @@ except ImportError:
 __all__ = ['perf_counter']
 
 
-from perf._utils import is_significant, python_implementation, python_has_jit  # noqa
-__all__.extend(('is_significant', 'python_implementation', 'python_has_jit'))
+from perf._utils import python_implementation, python_has_jit  # noqa
+__all__.extend(('python_implementation', 'python_has_jit'))
 
 from perf._metadata import format_metadata  # noqa
 __all__.append('format_metadata')